Fix detection of LeetCode V2 using improved document query logic (#52)
* Fix detection of LeetCode V2 using improved document query logic

Enhanced the function isLeetCodeV2 to more accurately determine if the current page is a version of LeetCode. The function now checks for the presence of specific root elements and body classes that indicate whether the page is using Chakra UI light or dark themes, as well as the presence of the "__next" root element. This ensures a more reliable and precise detection of LeetCode's version and UI themes, improving the overall functionality of the script.

Issue: #51

* Fixed submissionId for send solve